Feature/유저 인증 권한 추가

+++ cotobang/src/main/java/com/example/cotobang/service/CoinService.java
@@ -1,8 +1,10 @@
 package com.example.cotobang.service;
 
 import com.example.cotobang.domain.Coin;
+import com.example.cotobang.domain.User;
 import com.example.cotobang.dto.CoinDto;
 import com.example.cotobang.errors.CoinNotFoundException;
+import com.example.cotobang.errors.NotAuthorityException;
 import com.example.cotobang.respository.CoinRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -36,19 +38,24 @@ public class CoinService {
                 .orElseThrow(() -> new CoinNotFoundException(id));
     }
 
-    public Coin createCoin(CoinDto coinDto) {
+    public Coin createCoin(CoinDto coinDto, User user) {
         final Coin coin = Coin.builder()
                 .market(coinDto.getMarket())
                 .koreanName(coinDto.getKoreanName())
                 .englishName(coinDto.getEnglishName())
                 .description(coinDto.getDescription())
+                .user(user)
                 .build();
 
         return coinRepository.save(coin);
     }
 
-    public Coin updateCoin(Long targetId, CoinDto source) {
+    public Coin updateCoin(Long targetId, CoinDto source, User user) {
         Coin coin = getCoin(targetId);
+
+        if (!user.compare(coin.getUser())) {
+            throw new NotAuthorityException(user.getName());
+        }
 
         coin.change(
                 source.getMarket(),
@@ -60,8 +67,12 @@ public class CoinService {
         return coin;
     }
 
-    public Coin deleteCoin(Long id) {
+    public Coin deleteCoin(Long id, User user) {
         Coin coin = getCoin(id);
+
+        if (!user.compare(coin.getUser())) {
+            throw new NotAuthorityException(user.getName());
+        }
 
         coinRepository.delete(coin);
